requirements.txt
Django==2.2.6
django-crispy-forms==1.7.2
djangorestframework==3.10.3
gunicorn==19.9.0
mysqlclient==1.4.4
Pillow==6.2.0
pkg-resources==0.0.0
PyJWT==1.7.1
python-http-client==3.2.1
pytz==2019.2
sendgrid==6.1.0
sqlparse==0.3.0